backend/requirements.txt
asgiref==3.7.2
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
Werkzeug==2.3.7
